--- a/blurry/src/main/java/jp/wasabeef/blurry/Blur.java
+++ b/blurry/src/main/java/jp/wasabeef/blurry/Blur.java
@@ -6,6 +6,7 @@ import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.PorterDuff;
 import android.graphics.PorterDuffColorFilter;
+import android.os.Build;
 import android.renderscript.Allocation;
 import android.renderscript.Element;
 import android.renderscript.RSRuntimeException;