--- a/blurry/src/main/java/jp/wasabeef/blurry/BlurTask.java
+++ b/blurry/src/main/java/jp/wasabeef/blurry/BlurTask.java
@@ -10,6 +10,7 @@ import android.os.Looper;
 import android.util.Log;
 import android.view.PixelCopy;
 import android.view.View;
+import android.view.Window;
 
 import java.lang.ref.WeakReference;
 import java.util.concurrent.ExecutorService;
@@ -44,13 +45,16 @@ class BlurTask {
   private static final ExecutorService THREAD_POOL = Executors.newCachedThreadPool();
   private boolean bitmapExtractCompleted = false;
 
+  /**
+   * @param activity Nullable, will fall back to non-surface deprecated drawing-cache when no activity is supplied
+   */
   public BlurTask(Activity activity, View target, BlurFactor factor, Callback callback) {
     this.factor = factor;
     this.callback = callback;
     this.contextWeakRef = new WeakReference<>(target.getContext());
 
     long start = System.currentTimeMillis();
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && activity != null) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && activity != null && activity.getWindow().peekDecorView() != null) {
       // Use PixelCopy, PixelCopy copies from the surface, and can thus handle GoogleMap
       bitmap = Bitmap.createBitmap(target.getWidth(), target.getHeight(), Bitmap.Config.ARGB_8888);
 
@@ -58,14 +62,24 @@ class BlurTask {
       target.getLocationInWindow(locations);
       Rect rect = new Rect(locations[0], locations[1], locations[0] + target.getWidth(), locations[1] + target.getHeight());
 
-
-      PixelCopy.request(activity.getWindow(), rect, bitmap, copyResult -> {
-        if (copyResult != PixelCopy.SUCCESS) {
-          bitmap = extractBitmapByDeprecatedDrawingCache(target);
-        }
-        bitmapExtractCompleted = true;
-        execute();
-      }, new Handler(Looper.getMainLooper())); // We will get the callback on the handler, probably don't use main, maybe it has to be main if we want to extract bitmap old fashioned way on error
+      Window window = activity.getWindow();
+      // See javadoc on PixelCopy.request: https://developer.android.com/reference/android/view/PixelCopy#request(android.view.Window,%20android.graphics.Rect,%20android.graphics.Bitmap,%20android.view.PixelCopy.OnPixelCopyFinishedListener,%20android.os.Handler)
+      // - it requires that the window's decorView is already defined (handled in if-statement above)
+      // - and it requires that the window has a backing surface, they recommend postponing till after first onDraw. In this case catch error an fall back to deprecated drawing cache.
+      //   Alternatively we or the user must delay til after first Draw.
+      try {
+        PixelCopy.request(window, rect, bitmap, copyResult -> {
+          if (copyResult != PixelCopy.SUCCESS) {
+            bitmap = extractBitmapByDeprecatedDrawingCache(target);
+          }
+          bitmapExtractCompleted = true;
+          execute();
+        }, new Handler(Looper.getMainLooper())); // We will get the callback on the handler, probably don't use main, maybe it has to be main if we want to extract bitmap old fashioned way on error
+      } catch (IllegalArgumentException e) {
+        // Handle missing surface. See Android source-code https://github.com/aosp-mirror/platform_frameworks_base/blob/master/graphics/java/android/view/PixelCopy.java
+        // thus avoid IllegalArgumentException("Window doesn't have a backing surface!")
+        bitmap = extractBitmapByDeprecatedDrawingCache(target);
+      }
     } else {
       bitmap = extractBitmapByDeprecatedDrawingCache(target);
       bitmapExtractCompleted = true;