… README.md
… build.gradle
… example/build.gradle
… example/src/main/AndroidManifest.xml
… example/src/main/java/jp/wasabeef/example/blurry/MainActivity.kt
… gradle.properties