@arberg 2022-07-25
I've updated the code today, fixing a bug where the activity surface wasn't ready yet, and also cleaning up the new code a bit.